Develop aws

+++ database/sql/src/main/scala/cromwell/database/slick/tables/DockerHashStoreEntryComponent.scala
@@ -9,7 +9,7 @@ trait DockerHashStoreEntryComponent {
   import driver.api._
 
   class DockerHashStoreEntries(tag: Tag) extends Table[DockerHashStoreEntry](tag, "DOCKER_HASH_STORE_ENTRY") {
-    def dockerHashStoreEntryId = column[Int]("DOCKER_HASH_STORE_ENTRY_ID", O.PrimaryKey, O.AutoInc)
+    def dockerHashStoreEntryId = column[Long]("DOCKER_HASH_STORE_ENTRY_ID", O.PrimaryKey, O.AutoInc)
 
     def workflowExecutionUuid = column[String]("WORKFLOW_EXECUTION_UUID", O.Length(255))
 
@@ -19,20 +19,27 @@ trait DockerHashStoreEntryComponent {
 
     def dockerSize = column[Option[Long]]("DOCKER_SIZE", O.Default(None))
 
-    override def * = (workflowExecutionUuid, dockerTag, dockerHash, dockerSize, dockerHashStoreEntryId.?) <> (DockerHashStoreEntry.tupled, DockerHashStoreEntry.unapply)
+    override def * = (workflowExecutionUuid,
+                      dockerTag,
+                      dockerHash,
+                      dockerSize,
+                      dockerHashStoreEntryId.?
+    ) <> (DockerHashStoreEntry.tupled, DockerHashStoreEntry.unapply)
 
-    def ucDockerHashStoreEntryWeuDt = index("UC_DOCKER_HASH_STORE_ENTRY_WEU_DT", (workflowExecutionUuid, dockerTag), unique = true)
+    def ucDockerHashStoreEntryWeuDt =
+      index("UC_DOCKER_HASH_STORE_ENTRY_WEU_DT", (workflowExecutionUuid, dockerTag), unique = true)
   }
 
   val dockerHashStoreEntries = TableQuery[DockerHashStoreEntries]
 
-  val dockerHashStoreEntryIdsAutoInc = dockerHashStoreEntries returning dockerHashStoreEntries.map(_.dockerHashStoreEntryId)
+  val dockerHashStoreEntryIdsAutoInc =
+    dockerHashStoreEntries returning dockerHashStoreEntries.map(_.dockerHashStoreEntryId)
 
   /**
     * Useful for finding the docker hash store for a given workflow execution UUID
     */
-  val dockerHashStoreEntriesForWorkflowExecutionUuid = Compiled(
-    (workflowExecutionUuid: Rep[String]) => for {
+  val dockerHashStoreEntriesForWorkflowExecutionUuid = Compiled((workflowExecutionUuid: Rep[String]) =>
+    for {
       dockerHashStoreEntry <- dockerHashStoreEntries
       if dockerHashStoreEntry.workflowExecutionUuid === workflowExecutionUuid
     } yield dockerHashStoreEntry
