+++ database/sql/src/main/scala/cromwell/database/sql/tables/DockerHashStoreEntry.scala
@@ -1,10 +1,9 @@
 package cromwell.database.sql.tables
 
-case class DockerHashStoreEntry
-(
+case class DockerHashStoreEntry(
   workflowExecutionUuid: String,
   dockerTag: String,
   dockerHash: String,
   dockerSize: Option[Long],
-  dockerHashStoreEntryId: Option[Int] = None
+  dockerHashStoreEntryId: Option[Long] = None
 )
